Implement function for error reporting.
Create show_error() function to handle error reporting. The error
reporting mode can be set in the script or by setting the $DIALOG
environment variable. Modes currently implemented are:

    * (none specified) Report errors to the terminal using echo.
    * "kdialog" Report errors using a kdialog dialog.
    * "zenity" Report errors using a zenity dialog.

"zenity" is provided as the default mode because the script was
originally hard coded to always use zenity. We only want to add
options, not change existing behavior.